Move data fetching to the page level
